refactor: database connection for repository
