+++ packages/backend/src/utils/database/Database.ts
@@ -1,70 +1,22 @@
-import { DatabaseException } from '@/exceptions/DatabaseException';
-import { MySqlConnector } from '@/utils/database/connectors/MySqlConnector';
-import { RedisClient, RedisConnector } from '@/utils/database/connectors/RedisConnector';
+import { RedisClient } from '@/utils/database/connectors/RedisConnector';
 import { Pool } from 'mysql2/promise';
 
 export class Database {
 
-  private defaultConnector: MySqlConnector;
-  private cachingConnector: RedisConnector;
+  private readonly defaultDatabase: Pool;
+  private readonly cachingDatabase: RedisClient;
 
-  public async connect(): Promise<boolean> | never {
-    const expectedVariables = [
-      'MYSQL_HOST',
-      'MYSQL_PORT',
-      'MYSQL_USER',
-      'MYSQL_PASSWORD',
-      'MYSQL_DATABASE',
-      'REDIS_HOST',
-      'REDIS_PORT',
-    ];
-
-    expectedVariables.forEach((variable) => {
-      if (!process.env[variable]) {
-        throw new DatabaseException(`Please define ${variable} in .env`);
-      }
-    });
-
-    const {
-      MYSQL_HOST,
-      MYSQL_PORT,
-      MYSQL_USER,
-      MYSQL_PASSWORD,
-      MYSQL_DATABASE,
-      REDIS_HOST,
-      REDIS_PORT,
-      REDIS_USER,
-      REDIS_PASSWORD,
-    } = process.env;
-
-    this.defaultConnector = new MySqlConnector(
-      MYSQL_HOST,
-      Number(MYSQL_PORT),
-      MYSQL_USER,
-      MYSQL_PASSWORD,
-      MYSQL_DATABASE,
-    );
-    this.cachingConnector = new RedisConnector(
-      REDIS_HOST,
-      Number(REDIS_PORT),
-      REDIS_USER,
-      REDIS_PASSWORD,
-    );
-
-    await Promise.all([
-      this.defaultConnector.getConnection(),
-      this.cachingConnector.getConnection(),
-    ]);
-
-    return true;
+  public constructor(defaultDatabase: Pool, cachingDatabase) {
+    this.defaultDatabase = defaultDatabase;
+    this.cachingDatabase = cachingDatabase;
   }
 
-  public async getDefaultDatabase(): Promise<Pool> {
-    return this.defaultConnector.getConnection();
+  public query(sql: string, values: any) {
+    return this.defaultDatabase.query(sql, values);
   }
 
-  public async getCachingDatabase(): Promise<RedisClient> {
-    return this.cachingConnector.getConnection();
+  public execute(sql: string, values: any) {
+    return this.defaultDatabase.execute(sql, values);
   }
 
 }
